#1 Added webpack.dev ande webpack.prod configs
